+++ src/main/java/org/cicirello/ibp/LogRecord.java
@@ -57,13 +57,10 @@ final class LogRecord implements Serializable {
   }
 
   @Override
-  @SuppressWarnings("EqualsUnsafeCast")
   public boolean equals(Object other) {
-    /* // UNNECESSARY CHECK... access control prevents these cases
     if (other == null || !(other instanceof LogRecord)) {
-    	return false;
+      return false;
     }
-    */
     LogRecord o = (LogRecord) other;
     return type.equals(o.type) && data.equals(o.data) && timestamp == o.timestamp;
   }