+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+* LogRecord.equals(Object): correctly handle null and other types
 
 ### CI/CD
 * Integrated SpotBugs into build process.